backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
httptools==0.6.4
idna==3.10
numpy==2.2.6
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
